+++ package.json
@@ -22,6 +22,7 @@
     "@types/markdown-it": "^14.1.2",
     "lucide-react": "^0.536.0",
     "markdown-it": "^14.1.0",
+    "markdown-it-wikilinks": "^1.4.0",
     "pnpm": "^10.14.0",
     "prosemirror-example-setup": "^1.2.3",
     "prosemirror-markdown": "^1.13.2",
