+++ src/App.scss
@@ -9,7 +9,7 @@
 @use "./style/_file-tree.scss";
 @use "./style/_side-dock.scss";
 @use "./style/_title-bar.scss";
-@use "./style/_workspace-container.scss";
+@use "./style/_workspace.scss";
 
 :root {
   --green: #92e62c;
