+++ src/components/FileListing/index.ts
@@ -1,0 +1,2 @@
+export * from "./FileListing";
+export { default } from "./FileListing";
