+++ src/components/LeftLeaf/index.ts
@@ -1,0 +1,2 @@
+export * from "./LeftLeaf";
+export { default } from "./LeftLeaf"
