+++ src/components/MainPanel/index.ts
@@ -1,0 +1,2 @@
+export * from './MainPanel';
+export { default } from './MainPanel';
