+++ src/components/NoteEditor/NoteEditor.constants.ts
@@ -1,4 +1,16 @@
+import MarkdownIt from "markdown-it"
+import { MarkdownParser } from "prosemirror-markdown"
 import { MarkSpec, Schema } from "prosemirror-model"
+import { listIsTight } from "./NoteEditor.helpers"
+
+export const TYPING_TIMEOUT = 500;
+
+const MD = MarkdownIt("commonmark", {html: false})
+  // .use(require("markdown-it-wikilinks")({baseURL: "", relativeBaseURL: "", htmlAttributes: {
+  //   "class": "internal-link",
+  //   "rel": "nofollow",
+  // }}))
+
 
 /// Document schema for the data model used by CommonMark.
 export const markdownSchema: Schema<"blockquote" | "image" | "text" | "paragraph" | "code_block" | "doc" | "horizontal_rule" | "heading" | "ordered_list" | "bullet_list" | "list_item" | "hard_break" | "internal_link", "link" | "code" | "em" | "strong">
@@ -168,3 +180,40 @@ export const markdownSchema: Schema<"blockquote" | "image" | "text" | "paragraph
     }
   }
 })
+
+
+/// A parser parsing slightly extended [CommonMark](http://commonmark.org/),
+/// without inline HTML, with internal links, and producing a document in the basic schema.
+export const extendedMarkdownParser = new MarkdownParser(markdownSchema, MD, {
+  blockquote: {block: "blockquote"},
+  paragraph: {block: "paragraph"},
+  list_item: {block: "list_item"},
+  bullet_list: {block: "bullet_list", getAttrs: (_, tokens, i) => ({tight: listIsTight(tokens, i)})},
+  ordered_list: {block: "ordered_list", getAttrs: (tok, tokens, i) => ({
+    order: +tok.attrGet("start")! || 1,
+    tight: listIsTight(tokens, i)
+  })},
+  heading: {block: "heading", getAttrs: tok => ({level: +tok.tag.slice(1)})},
+  code_block: {block: "code_block", noCloseToken: true},
+  fence: {block: "code_block", getAttrs: tok => ({params: tok.info || ""}), noCloseToken: true},
+  hr: {node: "horizontal_rule"},
+  image: {node: "image", getAttrs: tok => ({
+    src: tok.attrGet("src"),
+    title: tok.attrGet("title") || null,
+    alt: tok.children![0] && tok.children![0].content || null
+  })},
+  hardbreak: {node: "hard_break"},
+  internal_link: {node: "internal_link", getAttrs: tok => ({
+    href: tok.attrGet("href"),
+    title: tok.attrGet("title") || null
+  })},
+
+  em: {mark: "em"},
+  strong: {mark: "strong"},
+  link: {mark: "link", getAttrs: tok => ({
+    href: tok.attrGet("href"),
+    title: tok.attrGet("title") || null
+  })},
+  code_inline: {mark: "code", noCloseToken: true}
+})
+
