+++ src/components/NoteEditor/index.ts
@@ -1,0 +1,2 @@
+export * from "./NoteEditor";
+export { default } from "./NoteEditor";
