+++ src/components/SettingsProvider/index.ts
@@ -1,0 +1,2 @@
+export * from "./SettingsProvider";
+export { default } from "./SettingsProvider";
