+++ src/components/SideDock/index.ts
@@ -1,0 +1,2 @@
+export * from "./SideDock";
+export { default } from "./SideDock"
