+++ src/components/TitleBar/index.ts
@@ -1,0 +1,2 @@
+export * from "./TitleBar";
+export { default } from "./TitleBar";
