+++ src/components/Workspace/index.ts
@@ -1,0 +1,2 @@
+export * from './Workspace';
+export { default } from './Workspace';
