--- src/components/text-editor-components/markdown-parser.ts
@@ -1,54 +1,0 @@
-/// add notes license notes for prosemirror-markdown
-
-import { MarkdownParser } from "prosemirror-markdown";
-import MarkdownIt from "markdown-it";
-import Token from "markdown-it/lib/token.mjs";
-import { markdownSchema } from "./markdown-schema";
-
-declare function require(name:string);
-
-function listIsTight(tokens: readonly Token[], i: number) {
-  while (++i < tokens.length)
-    if (tokens[i].type != "list_item_open") return tokens[i].hidden
-  return false
-}
-
-const md = MarkdownIt("commonmark", {html: false}).use(require("markdown-it-wikilinks")({baseURL: "", relativeBaseURL: "", htmlAttributes: {
-  "class": "internal-link",
-  "rel": "nofollow",
-}}))
-
-/// A parser parsing slightly extended [CommonMark](http://commonmark.org/),
-/// without inline HTML, with internal links, and producing a document in the basic schema.
-export const extendedMarkdownParser = new MarkdownParser(markdownSchema, md, {
-  blockquote: {block: "blockquote"},
-  paragraph: {block: "paragraph"},
-  list_item: {block: "list_item"},
-  bullet_list: {block: "bullet_list", getAttrs: (_, tokens, i) => ({tight: listIsTight(tokens, i)})},
-  ordered_list: {block: "ordered_list", getAttrs: (tok, tokens, i) => ({
-    order: +tok.attrGet("start")! || 1,
-    tight: listIsTight(tokens, i)
-  })},
-  heading: {block: "heading", getAttrs: tok => ({level: +tok.tag.slice(1)})},
-  code_block: {block: "code_block", noCloseToken: true},
-  fence: {block: "code_block", getAttrs: tok => ({params: tok.info || ""}), noCloseToken: true},
-  hr: {node: "horizontal_rule"},
-  image: {node: "image", getAttrs: tok => ({
-    src: tok.attrGet("src"),
-    title: tok.attrGet("title") || null,
-    alt: tok.children![0] && tok.children![0].content || null
-  })},
-  hardbreak: {node: "hard_break"},
-  internal_link: {node: "internal_link", getAttrs: tok => ({
-    href: tok.attrGet("href"),
-    title: tok.attrGet("title") || null
-  })},
-
-  em: {mark: "em"},
-  strong: {mark: "strong"},
-  link: {mark: "link", getAttrs: tok => ({
-    href: tok.attrGet("href"),
-    title: tok.attrGet("title") || null
-  })},
-  code_inline: {mark: "code", noCloseToken: true}
-})
