+++ src/main.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App";
-import SettingsProvider from "./state/SettingsContext";
-import TextEditor from "./components/TextEditor";
+import SettingsProvider from "@/components/SettingsProvider";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
