+++ src/style/_workspace.scss
@@ -1,4 +1,4 @@
-.workspace-container {
+.workspace {
   display: flex;
   flex-direction: row;
   flex: 1 1 0;
