+++ src/types.ts
@@ -1,0 +1,5 @@
+export type FileData = {
+  name: string;
+  filepath: string;
+  contents: string;
+}
